Syntax receivers renaming.
