+++ MongoExt.Generator/MongoExt.Generator.csproj
@@ -4,6 +4,7 @@
     <TargetFramework>netstandard2.0</TargetFramework>
     <LangVersion>latest</LangVersion>
     <Nullable>enable</Nullable>
+    <EnforceExtendedAnalyzerRules>true</EnforceExtendedAnalyzerRules>
     <Version>2.0.0</Version>
     <Authors>fedark</Authors>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
