+++ MongoExt.Generator/SyntaxReceivers/AggregatedSyntaxReceiver.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.CodeAnalysis;
+
+namespace MongoExt.Generator.SyntaxReceivers;
+
+internal class AggregatedSyntaxReceiver : ISyntaxReceiver
+{
+    public ModelSyntaxReceiver ModelSyntaxReceiver { get; } = new();
+    public CollectionDefinitionSyntaxReceiver CollectionDefinitionSyntaxReceiver { get; } = new();
+
+    public void OnVisitSyntaxNode(SyntaxNode syntaxNode)
+    {
+        ModelSyntaxReceiver.OnVisitSyntaxNode(syntaxNode);
+        CollectionDefinitionSyntaxReceiver.OnVisitSyntaxNode(syntaxNode);
+    }
+}
